Update 1toN probtrackx script, remove work_dir param

+++ subscripts/s3_probtrackx_alt.py
@@ -1,18 +1,17 @@
 #!/usr/bin/env python3
 from os.path import join,exists
-from subscripts.utilities import write,smart_remove,smart_mkdir
+from subscripts.utilities import write,smart_remove,smart_mkdir,get_edges_from_file
 from shutil import copyfile
 from parsl.app.app import python_app
 
 @python_app(executors=['s3'], cache=True)
-def s3_1_start(params, inputs=[]):
+def s3_1_start(params, volumes, inputs=[]):
     from subscripts.utilities import run,smart_remove,smart_mkdir,write,record_start,add_binary_vol,sub_binary_vol
     from os.path import join
     from shutil import copyfile
     sdir = params['sdir']
     use_gpu = params['use_gpu']
     stdout = params['stdout']
-    volumes = params['volumes']
     record_start(params)
     if use_gpu:
         write(stdout, "Running Probtrackx with GPU")
@@ -22,7 +21,6 @@ def s3_1_start(params, inputs=[]):
     smart_remove(outdir)
     smart_mkdir(outdir)
 
-    run_vtk = join(sdir, 'run_vtk.py')
     EDI_allvols = join(sdir,"EDI","allvols")
     allvoxelscortsubcort = join(sdir,"allvoxelscortsubcort.nii.gz")
     terminationmask = join(sdir,"terminationmask.nii.gz")
@@ -32,16 +30,19 @@ def s3_1_start(params, inputs=[]):
     copyfile(allvoxelscortsubcort, exclusion)
     copyfile(terminationmask, termination)
     for vol in volumes:
-        vol_file = join(EDI_allvols, vol + ".nii.gz")
+        vol_file = join(EDI_allvols, vol + "_s2fa.nii.gz")
+        if not exists(vol_file):
+            raise Exception('Failed to find volume {}'.format(vol_file))
         sub_binary_vol(vol_file, exclusion, params)
         add_binary_vol(vol_file, termination, params)
     run("fslmaths {} -add {} {}".format(exclusion, bs, exclusion), params)
 
-    run('/opt/vtk/bin/vtkpython {} {} {} {}'.format(run_vtk, termination, termination + '.png', 256), params)
-    run('/opt/vtk/bin/vtkpython {} {} {} {}'.format(run_vtk, exclusion, exclusion + '.png', 256), params)
+    # run_vtk = join(sdir, 'run_vtk.py')
+    # run('/opt/vtk/bin/vtkpython {} {} {} {}'.format(run_vtk, termination, termination + '.png', 256), params)
+    # run('/opt/vtk/bin/vtkpython {} {} {} {}'.format(run_vtk, exclusion, exclusion + '.png', 256), params)
 
 @python_app(executors=['s3'], cache=True)
-def s3_2_probtrackx(params, vol, inputs=[]):
+def s3_2_probtrackx(params, vol, volumes, inputs=[]):
     import time
     from subscripts.utilities import run,smart_remove,smart_mkdir,write,record_start,record_apptime,sub_binary_vol
     from os.path import join,exists,split
@@ -49,7 +50,6 @@ def s3_2_probtrackx(params, vol, inputs=[]):
     start_time = time.time()
     sdir = params['sdir']
     use_gpu = params['use_gpu']
-    volumes = params['volumes']
     pbtx_sample_count = int(params['pbtx_sample_count'])
     connectome_oneway = join(sdir, "connectome_oneway.dot")
     bedpostxResults = join(sdir,"bedpostx_b1000.bedpostX")
@@ -60,30 +60,38 @@ def s3_2_probtrackx(params, vol, inputs=[]):
     termination = join(outdir, "termination.nii.gz")
     EDI_allvols = join(sdir,"EDI","allvols")
 
-    vol_file = join(EDI_allvols, vol + ".nii.gz")
+    vol_file = join(EDI_allvols, vol + "_s2fa.nii.gz")
+    if not exists(vol_file):
+        raise Exception('Failed to find volume {}'.format(vol_file))
     vol_outdir = join(outdir, vol)
     smart_remove(vol_outdir)
     smart_mkdir(vol_outdir)
     waypoints = join(vol_outdir, 'waypoint.txt')
     for vol2 in volumes:
         if vol != vol2:
-            write(waypoints, join(EDI_allvols, vol2 + ".nii.gz"), params)
+            vol2_file = join(EDI_allvols, vol2 + "_s2fa.nii.gz")
+            if not exists(vol2_file):
+                raise Exception('Failed to find volume {}'.format(vol2_file))
+            write(waypoints, vol2_file, params)
     vol_termination = join(vol_outdir, "vol_termination.nii.gz")
     vol_exclusion = join(vol_outdir, "vol_exclusion.nii.gz")
     copyfile(termination, vol_termination)
     copyfile(exclusion, vol_exclusion)
     sub_binary_vol(vol_file, vol_termination, params)
     sub_binary_vol(vol_file, vol_exclusion, params)
+    vol_formatted = "fdt_paths.nii.gz"
 
     pbtx_args = (" -x {} ".format(vol_file) +
                 " --pd -l -c 0.2 -S 2000 --steplength=0.5 -P {}".format(pbtx_sample_count) +
                 " --waycond=OR --waypoints={}".format(waypoints) +
+                " --os2t --s2tastext --targetmasks={}".format(waypoints) +
                 " --stop={}".format(vol_termination) +
                 " --avoid={}".format(vol_exclusion) +
                 " --forcedir --opd" +
                 " -s {}".format(merged) +
                 " -m {}".format(nodif_brain_mask) +
-                " --dir={}".format(vol_outdir)
+                " --dir={}".format(vol_outdir) +
+                " --out={}".format(vol_formatted)
                 )
     if use_gpu:
         probtrackx2_sh = join(vol_outdir, "probtrackx2.sh")
@@ -95,17 +103,31 @@ def s3_2_probtrackx(params, vol, inputs=[]):
     else:
         run("probtrackx2" + pbtx_args, params)
 
+    # vol_connectome = join(vol_outdir,"connectome.dot")
+    # waytotal = join(vol_outdir, "waytotal")
+    # if not exists(waytotal):
+    #     write(stdout, 'Error: failed to find waytotal for volume {}'.format(vol))
+
+    # with open(waytotal, 'r') as f:
+    #     waytotal_count = f.read().strip()
+    #     fdt_count = run("fslmeants -i {} -m {} | head -n 1".format(join(vol_outdir, vol_formatted), vol_file), params)
+    #     if not is_float(waytotal_count):
+    #         raise Exception("Failed to read waytotal_count value {}".format(waytotal_count))
+    #     if not is_float(fdt_count):
+    #         raise Exception("Failed to read fdt_count value {}".format(fdt_count))
+    #     write(vol_connectome, "{} {} {}".format(vol, waytotal_count, fdt_count))
+
     record_apptime(params, start_time, 1)
 
 @python_app(executors=['s3'], cache=True)
-def s3_3_combine(params, inputs=[]):
+def s3_3_combine(params, volumes, inputs=[]):
     import time
     from subscripts.utilities import run,record_apptime,record_finish,update_permissions,write
     from os.path import join,exists
     sdir = params['sdir']
-    volumes = params['volumes']
+    # volumes = params['volumes']
     start_time = time.time()
-    run_vtk = join(sdir, 'run_vtk.py')
+    # run_vtk = join(sdir, 'run_vtk.py')
     outdir = join(sdir, 'fast_outdir')
     allvoxelscortsubcort = join(sdir,"allvoxelscortsubcort.nii.gz")
     total = join(outdir, 'FAtractsumsTwoway.nii.gz')
@@ -115,19 +137,25 @@ def s3_3_combine(params, inputs=[]):
         pbtx_result = join(vol_outdir, 'fdt_paths.nii.gz')
         run("fslmaths {} -thrP 5 -bin {}".format(pbtx_result, pbtx_result), params)
         run("fslmaths {} -add {} {}".format(pbtx_result, total, total), params)
-    run('/opt/vtk/bin/vtkpython {} {} {} {}'.format(run_vtk, total, total + '.png', 256), params)
+
+        waytotal = join(vol_outdir, "waytotal")
+    # run('/opt/vtk/bin/vtkpython {} {} {} {}'.format(run_vtk, total, total + '.png', 256), params)
 
     update_permissions(params)
     record_apptime(params, start_time, 2)
     record_finish(params)
 
-def setup_s3(params, inputs):
+def setup_s3_alt(params, inputs):
     sdir = params['sdir']
-    volumes = params['volumes']
-    run_vtk = join(sdir, 'run_vtk.py')
-    copyfile('subscripts/run_vtk.py', run_vtk)
-    s3_1_future = s3_1_start(params, inputs=inputs)
+    pbtx_edge_list = params['pbtx_edge_list']
+    volumes = set([])
+    for edge in get_edges_from_file(pbtx_edge_list):
+        volumes.add(edge[0])
+        volumes.add(edge[1])
+    # run_vtk = join(sdir, 'run_vtk.py')
+    # copyfile('subscripts/run_vtk.py', run_vtk)
+    s3_1_future = s3_1_start(params, volumes, inputs=inputs)
     s3_2_futures = []
     for vol in volumes:
-        s3_2_futures.append(s3_2_probtrackx(params, vol, inputs=[s3_1_future]))
-    return s3_3_combine(params, inputs=s3_2_futures)
+        s3_2_futures.append(s3_2_probtrackx(params, vol, volumes, inputs=[s3_1_future]))
+    return s3_3_combine(params, volumes, inputs=s3_2_futures)
